Inheritance
In this exercise, your task is to define a new class which will be inherited from the class Customer defined in the chapter's theory part.




01  class Customer:
02      name = "John Johnsson"
03      total = 1000
04      paymenttype = "Masterexpress"
05      number = "1234-5678-9012345"
06
07      def printout(self):
08          print("Name: ", self.name)
09          print("Total: ", self.total)
10          print("Payment type: ", self.paymenttype)
11          print("Card/Bill number: ", self.number)


Define a new class called ManageCustomer which has the following abilities:
1) A method "addbill", which increases the attribute total by 50, and
2) a method "payment" which decreases the attribute total by 100.


Remember that you only need to write the new inheriting class ManageCustomer. The system will test the class, but in the submitted answer only the class definition is needed. If the class is defined correctly, the program will print the following:




>>> 
Name:  Homer Griffin
Total:  850
Payment type:  Masterexpress
Card/Bill number:  1234-5678-9012345
>>>